Change license to MIT to allow inclusion into NodeCore
For any code, images, etc. to be possible to include in upstream
NodeCore, the project contribution guidelines need to be followed,
which includes using the required license:

https://gitlab.com/sztest/nodecore/-/blob/master/CONTRIBUTING

By accepting and merging this pull request, WintersKnight94 will
relicense this project under the MIT license (the old license can
still apply to previously-released versions as well) and agree
to the terms of the contribution guidelines.